refactor(onboarding): move onboardings to separate component
