+++ src/scripts/components/main/app/app.styl
@@ -39,12 +39,6 @@
   z-index: 4
   fill: $textWhite
 
-.welcomeOverlay
-  background: linear-gradient(90deg, rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
-
-.onboardingOverlay
-  background: transparent
-
 @media screen and (max-width: 480px)
   :global(path.hideText)
     display: none
@@ -52,6 +46,3 @@
   .logo
     overflow: hidden
     width: 150px
-
-  .welcomeOverlay
-    background: linear-gradient(rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
