+++ src/scripts/components/main/app/app.tsx
@@ -1,4 +1,4 @@
-import React, {FunctionComponent, useState} from 'react';
+import React, {FunctionComponent} from 'react';
 import {Provider as StoreProvider, useSelector} from 'react-redux';
 import {IntlProvider} from 'react-intl';
 import {HashRouter as Router, Switch, Route, Link} from 'react-router-dom';
@@ -22,11 +22,7 @@ import ShowcaseSelector from '../../stories/showcase-selector/showcase-selector'
 import DataViewer from '../data-viewer/data-viewer';
 import Tracking from '../tracking/tracking';
 import AboutProjectOverlay from '../about-project-overlay/about-project-overlay';
-import WelcomeScreen from '../welcome-screen/welcome-screen';
-import Overlay from '../overlay/overlay';
-import Onboarding from '../onboarding/onboarding';
-import {welcomeScreenSelector} from '../../../selectors/welcome-screen';
-
+import Onboarding from '../welcome/welcome';
 import translations from '../../../i18n';
 import {useStoryMarkers} from '../../../hooks/use-story-markers';
 
@@ -54,8 +50,6 @@ const App: FunctionComponent = () => (
 const TranslatedApp: FunctionComponent = () => {
   const markers = useStoryMarkers();
   const language = useSelector(languageSelector);
-  const showWelcomeScreen = useSelector(welcomeScreenSelector);
-  const [onboardingStep, setOnboardingStep] = useState<number | null>(null);
 
   const logo = (
     <Link to="/about">
@@ -71,25 +65,7 @@ const TranslatedApp: FunctionComponent = () => {
         <Switch>
           <Route path="/" exact>
             {logo}
-            {showWelcomeScreen ? (
-              <Overlay
-                className={styles.welcomeOverlay}
-                showCloseButton={false}>
-                <WelcomeScreen onStartOnboarding={() => setOnboardingStep(1)} />
-              </Overlay>
-            ) : (
-              onboardingStep && (
-                <Overlay
-                  className={styles.onboardingOverlay}
-                  showCloseButton={false}>
-                  <Onboarding
-                    step={onboardingStep}
-                    onPageChange={(step: number) => setOnboardingStep(step)}
-                    onClose={() => setOnboardingStep(null)}
-                  />
-                </Overlay>
-              )
-            )}
+            <Onboarding />
             <DataViewer markers={markers} backgroundColor={'#10161A'} />
             <Navigation />
             <TimeSlider />
