+++ src/scripts/components/main/welcome/welcome.styl
@@ -1,0 +1,10 @@
+
+.welcomeOverlay
+  background: linear-gradient(90deg, rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
+
+.onboardingOverlay
+  background: transparent
+
+@media screen and (max-width: 480px)
+  .welcomeOverlay
+    background: linear-gradient(rgba(16, 22, 26, 0.85) 0%, rgba(16, 22, 26, 0.81) 48%, rgba(16, 22, 26, 0.79) 50%, rgba(16, 22, 26, 0.75) 52%, rgba(16, 22, 26, 0.2) 60%, rgba(16, 22, 26, 0.1) 70%, rgba(16, 22, 26, 0) 100%)
